Rename betty/**/*/test___init__.py to betty/**/test_*.py
