+++ betty/tests/locale/test_localizable.py
