+++ betty/tests/service/plugin/definition/test_collection.py
@@ -12,7 +12,7 @@ from betty.test_utils.plugin import (
     DummyPluginDefinition,
     DummyPluginWithLifeCycle,
 )
-from betty.tests.service.plugin.test___init__ import (
+from betty.tests.service.test_plugin import (
     PluginServiceManagerTestBase,
 )
