+++ betty/tests/service/plugin/instance/test_single.py
@@ -8,7 +8,7 @@ from betty.test_utils.plugin import (
     DummyPluginManufacturer,
     DummyPluginWithLifeCycle,
 )
-from betty.tests.service.plugin.test___init__ import (
+from betty.tests.service.test_plugin import (
     PluginServiceManagerTestBase,
 )
