+++ betty/tests/service/plugin/test_instance.py
@@ -18,7 +18,7 @@ from betty.test_utils.plugin import (
     DummyPluginTwo,
     DummyPluginWithLifeCycle,
 )
-from betty.tests.service.plugin.test___init__ import (
+from betty.tests.service.test_plugin import (
     PluginServiceManagerTestBase,
 )
