+++ betty/tests/test_console.py
