+++ betty/tests/test_entity.py
