+++ betty/tests/test_indicator.py
