+++ betty/tests/test_life_cycle.py
