+++ betty/tests/test_plugin.py
