+++ betty/tests/test_rich.py
